% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData.R
\name{genData}
\alias{genData}
\title{Generate a simulated dataset}
\usage{
genData(N, sce = "VCAR1")
}
\arguments{
\item{N}{Integer value: number of observations that will be generated}

\item{sce}{Character value: The scenario number of simulation studies in Sun et al. (2020+).
Possible values are "VCAR1", "VAR2", "VAR3", "VCAR4", "VAR-M5", "VAR-M6", "VNAR7", "VNAR8", "VNAR9", "VNAR10".}
}
\value{
A list of simulated data, contains the three data frames that will be used in the function reVAR
\item{nonEventDF1_s2}{A data frame that is used for fitting the non-event time model, contains ID, event time information and covariates.
Each row records the covariates on a time interval.
\code{ID} is the subject ID for each observation;
\code{Start} is the starting time of the interval;
\code{End} is the ending time of the interval;
\code{Status} is the status indicator, 1 = non-event visit, 0 = no;
\code{Status2} is the status indicator, 1 = event visit, 0 = no.
The other columns are covariate values for the non-event model on the time interval (\code{Start}, \code{End}].}


\item{nonEventDF2}{A data frame that contains the ID, non-event visit times and covariates (for the event model) measured at each non-event visit.
\code{ID} is the subject ID for each observation;
\code{Time} are time zero and the non-event visit times.
The other columns are covariate values measured at \code{Time}.}


\item{eventDF}{A data frame that contains the ID, event visit times and covariates (for the event model) measured at each event visit.
\code{ID} is the subject ID for each observation;
\code{Time} are time zero and the event times.
The other columns are covariate values measured at \code{Time}.}
}
\description{
This function generates a simulated dataset under scenarios of the
simulation studies in Sun et al. (2020+).
}
\examples{
set.seed(1)
dat <- genData(100)
str(dat)
}
