% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reVARBoot.R
\name{reVARBoot}
\alias{reVARBoot}
\title{Resampling for variance estimation using reVAR}
\usage{
reVARBoot(nonEventDF1, nonEventDF2, eventDF, tau, h, B)
}
\arguments{
\item{nonEventDF1}{A data  frame that is used for fitting the non-event time model, contains ID, event time information and covariates}

\item{nonEventDF2}{A dataframe that contains the ID, non-event visit times and covariates measured at each non-event visit}

\item{eventDF}{A data frame that contains the ID, event visit times and covariates measured at each event visit}

\item{tau}{Data on [0,tau] will be used in the estimating equation in the second step}

\item{h}{Bandwidth for kernel estimation}

\item{B}{Number of bootstrap samples}
}
\value{
A matrix of estimates from bootstrapped samples
}
\description{
This function gives the estimates using bootstrapped samples.
}
\examples{
set.seed(1)
dat <- genData(100)
estBoot <- reVARBoot(dat$nonEventDF1,dat$nonEventDF2,dat$eventDF,4.5, h = 2*100^(-1/3), B = 500)
# Variance Estimation
apply(estBoot, 2, sd)
}
