% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reVCAR.R
\name{reVCAR}
\alias{reVCAR}
\title{Fit the proportional rate model with intermittently measured time-dependent covariates under VCAR}
\usage{
reVCAR(nonEventDF2, eventDF, tau, h)
}
\arguments{
\item{nonEventDF2}{A data frame that contains the ID, non-event visit times and covariates (for the event model) measured at each non-event visit.
The first two columns must be \code{ID} and \code{Time}:
\code{ID} is the subject ID for each observation;
\code{Time} include time zero and the non-event visit times.
The other columns are covariate values measured at \code{Time}.}

\item{eventDF}{A data frame that contains the ID, event visit times and covariates (for the event model) measured at each event visit.
The first two columns must be \code{ID} and \code{Time}:
\code{ID} is the subject ID for each observation;
\code{Time} include time zero and the event times.
The other columns are covariate values measured at \code{Time}.}

\item{tau}{Numeric value: data on [0,\code{tau}] will be used in the estimating equation in the second step.}

\item{h}{Numeric value: bandwidth for kernel estimation.}
}
\value{
A vector of the coefficients
}
\description{
This function fit the proportional rate model with intermittently
measured time-dependent covariates under the visiting completely at random (VCAR) assumption.
}
\examples{
set.seed(1)
dat <- genData(100)
ndf2 <- dat$nonEventDF2
edf <- dat$eventDF
fit <- reVCAR(ndf2, edf, tau = 4.5, h = 0.4)
}
