% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reVCARBoot.R
\name{reVCARBoot}
\alias{reVCARBoot}
\title{Resampling for variance estimation using reVCAR}
\usage{
reVCARBoot(nonEventDF2, eventDF, tau, h, B)
}
\arguments{
\item{nonEventDF2}{A data frame that contains the ID, non-event visit times and covariates (for the event model) measured at each non-event visit.
The first two columns must be \code{ID} and \code{Time}:
\code{ID} is the subject ID for each observation;
\code{Time} include time zero and the non-event visit times.
The other columns are covariate values measured at \code{Time}.}

\item{eventDF}{A data frame that contains the ID, event visit times and covariates (for the event model) measured at each event visit.
The first two columns must be \code{ID} and \code{Time}:
\code{ID} is the subject ID for each observation;
\code{Time} include time zero and the event times.
The other columns are covariate values measured at \code{Time}.}

\item{tau}{Numeric value: data on [0,\code{tau}] will be used in the estimating equation in the second step.}

\item{h}{Numeric value: bandwidth for kernel estimation.}

\item{B}{Integer value: number of bootstrap samples.}
}
\value{
A matrix of the coefficients from the bootstrap samples;
each row corresponds to the estimate on a bootstrapp sample
}
\description{
This function gives the estimates using bootstrapped samples.
}
\examples{
set.seed(1)
dat <- genData(100)
ndf2 <- dat$nonEventDF2
edf <- dat$eventDF
fit <- reVCAR(ndf2, edf, tau = 4.5, h = 0.4)
bt <- reVCARBoot(ndf2, edf, tau = 4.5, h = 0.4, B = 500)
# Variance Estimation
apply(bt, 2, sd)
}
