% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGen.R
\name{genData}
\alias{genData}
\title{Generate a simulated dataset}
\usage{
genData(N)
}
\arguments{
\item{N}{Number of observations that will be generated}
}
\value{
A list of simulated data, contains the data frames that will be used in the function reVAR
}
\description{
This function generate a simulated dataset under the first scenario in the
simulation study of the paper.
}
\examples{
set.seed(1)
dat <- genData(100)
}
