% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reVAR.R
\name{reVAR}
\alias{reVAR}
\title{Fit the proportional rate model with intermittently measured time-dependent covariates under VAR}
\usage{
reVAR(nonEventDF1, nonEventDF2, eventDF, tau, h)
}
\arguments{
\item{nonEventDF1}{A data frame that is used for fitting the non-event time model, contains ID, event time information and covariates}

\item{nonEventDF2}{A data frame that contains the ID, non-event visit times and covariates measured at each non-event visit}

\item{eventDF}{A data frame that contains the ID, event visit times and covariates measured at each event visit}

\item{tau}{Data on [0,tau] will be used in the estimating equation in the second step}

\item{h}{Bandwidth for kernel estimation}
}
\value{
A list of parameter estimates in the first and second step
}
\description{
This function fit the proportional rate model with intermittently
measured time-dependent covariates under VAR.
}
\examples{
set.seed(1)
dat <- genData(100)
fit <- reVAR(dat$nonEventDF1,dat$nonEventDF2,dat$eventDF, tau = 4.5, h = 2*100^(-1/3))
}
